decompose createContext into smaller files
