+++ invoice-backend/eslint.config.mjs
@@ -2,7 +2,7 @@ import eslint from "@eslint/js";
 import { globalIgnores } from "eslint/config";
 import tseslint from "typescript-eslint";
 import eslintPluginPrettierRecommended from "eslint-plugin-prettier/recommended";
-import pluginSecurity from "eslint-plugin-security";
+// import pluginSecurity from "eslint-plugin-security";
 
 // /** @type {import('eslint').Linter.Config[]} */
 export default tseslint.config(
@@ -25,9 +25,9 @@ export default tseslint.config(
   },
   eslint.configs.recommended,
   ...tseslint.configs.recommendedTypeChecked,
-  tseslint.configs.strict,
+  // tseslint.configs.strict,
   eslintPluginPrettierRecommended,
-  pluginSecurity.configs.recommended,
+  // pluginSecurity.configs.recommended,
   {
     rules: {
       "@typescript-eslint/no-unnecessary-condition": "error",
